+++ src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import Announcement from "../components/Announcement";
 import QuickCategories from "../components/QuickCategories";
 import Footer from "../components/Footer";
@@ -7,9 +7,24 @@ import Newsletter from "../components/Newsletter";
 import Products from "../components/Products";
 import Slider from "./../components/Slider";
 import { Alerta } from "../components/Alerts";
+import apiBase from "../api/apiBase";
 
 const Home = (props) => {
   const {productosEnCarrito, setProductoSeleccionado, sumarAlCarrito, setCategoriaSeleccionada} = props;
+  const [products, setProducts] = useState([]);
+
+  useEffect(() => {    
+    const fetchRandomProducts = async () => {
+      try{
+        const response = await apiBase.get("/productos/random/4")
+        setProducts(response.data)
+      } catch (err){
+        console.log("ayayay")
+      }
+    }
+
+    fetchRandomProducts();
+  }, []);
   
   return (
     <div>
@@ -23,6 +38,7 @@ const Home = (props) => {
       <Products 
         setProductoSeleccionado={setProductoSeleccionado} 
         sumarAlCarrito={sumarAlCarrito}
+        products={products}
       />
       <Newsletter/>
       <Footer/>
